requirements.txt
datashader>=0.10.0
leidenalg>=0.8.3
networkit>=7.0
networkx>=2.3
pandas>=0.25.3
plotly>=3.10.0
python-igraph>=0.8.3
python-louvain>=0.14